+++ packages/dsw-mailer/dsw/mailer/__init__.py
@@ -1,3 +1,4 @@
 from .cli import main
+from .handlers import lambda_handler
 
-__all__ = ['main']
+__all__ = ['main', 'lambda_handler']