+++ packages/dsw-mailer/dsw/mailer/mailer.py
@@ -66,11 +66,11 @@ class Mailer(CommandWorker):
             built_at=built_at,
         )
 
-    def run(self):
+    def _run_preparation(self) -> CommandQueue:
         Context.get().app.db.connect()
         # prepare
         self._update_component_info()
-        # work in queue
+        # init queue
         LOG.info('Preparing command queue')
         queue = CommandQueue(
             worker=self,
@@ -79,7 +79,17 @@ class Mailer(CommandWorker):
             component=CMD_COMPONENT,
             timeout=Context.get().app.cfg.db.queue_timout,
         )
+        return queue
+
+    def run(self):
+        LOG.info('Starting mailer worker (loop)')
+        queue = self._run_preparation()
         queue.run()
+
+    def run_once(self):
+        LOG.info('Starting mailer worker (once)')
+        queue = self._run_preparation()
+        queue.run_once()
 
     def work(self, cmd: PersistentCommand):
         # update Sentry info
@@ -96,7 +106,8 @@ class Mailer(CommandWorker):
         # get tenant config from DB
         tenant_cfg = app_ctx.db.get_tenant_config(tenant_uuid=cmd.tenant_uuid)
         LOG.debug(f'Tenant config from DB: {tenant_cfg}')
-        rq.style.from_dict(tenant_cfg.look_and_feel)
+        if tenant_cfg is not None:
+            rq.style.from_dict(tenant_cfg.look_and_feel)
         # get mailer config from DB
         mail_cfg = _transform_mail_config(
             cfg=app_ctx.db.get_mail_config(tenant_uuid=cmd.tenant_uuid),