+++ packages/dsw-data-seeder/CHANGELOG.md
@@ -7,6 +7,10 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
 
+### Added
+
+- Support for Lambda deployment
+
 
 ## [4.2.1]